+++ tests/helpers/random_int.h
@@ -28,7 +28,7 @@ namespace simdutf::tests::helpers {
   public:
     RandomIntRanges(std::initializer_list<std::pair<uint64_t, uint64_t>> ranges, uint64_t seed);
 
-    uint64_t operator()();
+    uint32_t operator()();
   };
 
 }
